+++ viadot/utils.py
@@ -5,7 +5,8 @@ import pandas as pd
 import prefect
 import pyodbc
 import requests
-from prefect.utilities.graphql import EnumValue, with_args
+
+# from prefect.utilities.graphql import EnumValue, with_args
 from requests.adapters import HTTPAdapter
 from requests.exceptions import ConnectionError, HTTPError, ReadTimeout, Timeout
 from requests.packages.urllib3.util.retry import Retry